extracted generic tag metadata class
